test: fix and refactor broken tests
