+++ internal/k8s/fake/client.go
@@ -1,6 +1,6 @@
-// Package mock contains utilities intended for mocking out functionality in test code.
+// Package fake contains utilities intended for mocking out functionality in test code.
 // Most notably, it contains mocks for interacting with the Kubernetes API.
-package mock
+package fake
 
 import (
 	"github.com/stretchr/testify/mock"
@@ -9,35 +9,35 @@ import (
 	"k8s.io/client-go/rest"
 )
 
-// KubeClient is a mock k8s.KubeClient for interacting with a fake Kubernetes API, primarily intended for unit testing.
-type KubeClient struct {
+// Client is a fake k8s.Client for interacting with a fake Kubernetes API, primarily intended for unit testing.
+type Client struct {
 	mock.Mock
 }
 
 // ListJobs returns a list of jobs based on the Jobs field in KubeClient.
-func (client *KubeClient) ListJobs() (*batchv1.JobList, error) {
+func (client *Client) ListJobs() ([]batchv1.Job, error) {
 	args := client.Called()
-	jobs, _ := args.Get(0).(*batchv1.JobList)
+	jobs, _ := args.Get(0).([]batchv1.Job)
 
 	return jobs, args.Error(1)
 }
 
 // CreateJob simulates creating a job.
-func (client *KubeClient) CreateJob(job *batchv1.Job) error {
+func (client *Client) CreateJob(job *batchv1.Job) error {
 	args := client.Called(job)
 
 	return args.Error(0)
 }
 
 // DeleteJob simulates deleting a job.
-func (client *KubeClient) DeleteJob(name string) error {
+func (client *Client) DeleteJob(name string) error {
 	args := client.Called(name)
 
 	return args.Error(0)
 }
 
 // GetJob searches through the items in the Jobs field in KubeClient, returning the first item with a matching name.
-func (client *KubeClient) GetJob(name string) (*batchv1.Job, error) {
+func (client *Client) GetJob(name string) (*batchv1.Job, error) {
 	args := client.Called(name)
 	job, _ := args.Get(0).(*batchv1.Job)
 
@@ -45,7 +45,7 @@ func (client *KubeClient) GetJob(name string) (*batchv1.Job, error) {
 }
 
 // GetJobLogs simulates returning a rest.Request that will stream logs for a the job with the matching name.
-func (client *KubeClient) GetJobLogs(name string, opts *corev1.PodLogOptions) (*rest.Request, error) {
+func (client *Client) GetJobLogs(name string, opts *corev1.PodLogOptions) (*rest.Request, error) {
 	args := client.Called(name, opts)
 	req, _ := args.Get(0).(*rest.Request)
