+++ internal/k8s/fake/jobparser.go
@@ -1,4 +1,4 @@
-package mock
+package fake
 
 import (
 	"github.com/stretchr/testify/mock"
